Move project to IntelliJ
