+++ .idea/artifacts/gabriel.xml
@@ -1,0 +1,11 @@
+<component name="ArtifactManager">
+  <artifact type="jar" name="gabriel">
+    <output-path>$PROJECT_DIR$/out/artifacts/gabriel</output-path>
+    <root id="archive" name="gabriel.jar">
+      <element id="directory" name="META-INF">
+        <element id="file-copy" path="$PROJECT_DIR$/src/main/java/META-INF/MANIFEST.MF" />
+      </element>
+      <element id="module-output" name="gabriel" />
+    </root>
+  </artifact>
+</component>
