+++ .idea/dictionaries/ozan.xml
@@ -1,0 +1,7 @@
+<component name="ProjectDictionaryState">
+  <dictionary name="ozan">
+    <words>
+      <w>introducer</w>
+    </words>
+  </dictionary>
+</component>
