--- src/gabriel/Controller/Hasher.java
@@ -1,17 +1,0 @@
-package gabriel.Controller;
-
-import java.io.FileNotFoundException;
-import java.io.IOException;
-import java.io.RandomAccessFile;
-import java.security.MessageDigest;
-import java.security.NoSuchAlgorithmException;
-
-public class Hasher {
-
-    public static String calculateFileHash(String path) throws IOException, NoSuchAlgorithmException {
-        
-        
-        
-        return null;
-    }
-}
