--- src/gabriel/Controller/Receiver.java
@@ -1,23 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package gabriel.Controller;
-
-/**
- *
- * @author EmreDan
- * This class is to Receive Messages and Resources from other Connections
- */
-public class Receiver implements Runnable{
-    
-    
-    @Override
-    public void run() {
-        
-        
-        
-    }
-    
-}
