--- src/gabriel/Controller/Sender.java
@@ -1,20 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package gabriel.Controller;
-
-/**
- *
- * @author EmreDan
- * This class is to send Messages and Resources
- */
-public class Sender implements Runnable {
-
-    @Override
-    public void run() {
-        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
-    }
-    
-}
