--- src/gabriel/MainScreen.java
@@ -1,324 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package gabriel;
-
-import gabriel.Controller.Introducer;
-import gabriel.models.Connection;
-import gabriel.models.Index;
-import gabriel.models.Node;
-import java.awt.HeadlessException;
-import java.io.BufferedInputStream;
-import java.io.IOException;
-import java.io.ObjectInputStream;
-import java.net.Inet4Address;
-import java.net.ServerSocket;
-import java.net.Socket;
-import java.util.ArrayList;
-import java.util.concurrent.Executors;
-import java.util.logging.Level;
-import java.util.logging.Logger;
-import javax.swing.JOptionPane;
-
-/**
- *
- * @author TheDoctor
- */
-public class MainScreen extends javax.swing.JFrame {
-
-    /**
-     * Creates new form MainScreen
-     */
-    Gabriel gabriel;
-    Introducer mainIntroducer;
-    ConnectionManager connectionManager;
-
-    public MainScreen() {
-        initComponents();
-        gabriel = new Gabriel();
-        connectionManager = gabriel.getConnectionManager();
-        mainIntroducer = gabriel.getMainIntroducer();
-        mainIntroducer.checkHostsBruteForce(gabriel.getSubnet());
-        mainIntroducer.getHostCheckerAll().getHostIps().forEach((node) -> {
-            IpListBox.add(node);
-        });
-
-    }
-
-    /**
-     * This method is called from within the constructor to initialize the form.
-     * WARNING: Do NOT modify this code. The content of this method is always
-     * regenerated by the Form Editor.
-     */
-    @SuppressWarnings("unchecked")
-    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
-    private void initComponents() {
-
-        ScanNetworkButton = new javax.swing.JButton();
-        ConnectButton = new javax.swing.JButton();
-        jLabel1 = new javax.swing.JLabel();
-        jLabel2 = new javax.swing.JLabel();
-        ConnectionListBox = new java.awt.List();
-        jLabel3 = new javax.swing.JLabel();
-        IpListBox = new java.awt.List();
-        jLabel4 = new javax.swing.JLabel();
-        filler1 = new javax.swing.Box.Filler(new java.awt.Dimension(0, 0), new java.awt.Dimension(0, 0), new java.awt.Dimension(0, 0));
-        listenConnection = new javax.swing.JButton();
-        jLabel5 = new javax.swing.JLabel();
-        NetworkIndexListBox = new java.awt.List();
-        LocalIndexListBox = new java.awt.List();
-
-        setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
-
-        ScanNetworkButton.setText("Refresh Network List");
-        ScanNetworkButton.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                ScanNetworkButtonActionPerformed(evt);
-            }
-        });
-
-        ConnectButton.setText("Connect");
-        ConnectButton.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                ConnectButtonActionPerformed(evt);
-            }
-        });
-
-        jLabel1.setFont(new java.awt.Font("Sitka Text", 1, 14)); // NOI18N
-        jLabel1.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        jLabel1.setText("Nodes in your AdHoc Network");
-
-        jLabel2.setFont(new java.awt.Font("Tahoma", 1, 14)); // NOI18N
-        jLabel2.setText("Welcome to Organic AdHoc Networks!");
-
-        jLabel3.setFont(new java.awt.Font("Sitka Text", 1, 14)); // NOI18N
-        jLabel3.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        jLabel3.setText("Active Connections");
-        jLabel3.setToolTipText("");
-
-        IpListBox.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                IpListBoxActionPerformed(evt);
-            }
-        });
-
-        jLabel4.setText("Your File Index Table");
-
-        listenConnection.setText("Get Index");
-        listenConnection.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                listenConnectionActionPerformed(evt);
-            }
-        });
-
-        jLabel5.setText("Neighbour Index Table");
-
-        javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
-        getContentPane().setLayout(layout);
-        layout.setHorizontalGroup(
-            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(layout.createSequentialGroup()
-                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addGroup(layout.createSequentialGroup()
-                        .addGap(10, 10, 10)
-                        .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                            .addGroup(layout.createSequentialGroup()
-                                .addGap(174, 174, 174)
-                                .addComponent(ConnectButton, javax.swing.GroupLayout.PREFERRED_SIZE, 83, javax.swing.GroupLayout.PREFERRED_SIZE)
-                                .addGap(0, 0, Short.MAX_VALUE))
-                            .addGroup(layout.createSequentialGroup()
-                                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                                    .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, layout.createSequentialGroup()
-                                        .addGap(0, 18, Short.MAX_VALUE)
-                                        .addComponent(jLabel1, javax.swing.GroupLayout.PREFERRED_SIZE, 279, javax.swing.GroupLayout.PREFERRED_SIZE))
-                                    .addGroup(layout.createSequentialGroup()
-                                        .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING, false)
-                                            .addComponent(IpListBox, javax.swing.GroupLayout.Alignment.LEADING, javax.swing.GroupLayout.DEFAULT_SIZE, 269, Short.MAX_VALUE)
-                                            .addComponent(jLabel4, javax.swing.GroupLayout.Alignment.LEADING)
-                                            .addComponent(ScanNetworkButton, javax.swing.GroupLayout.Alignment.LEADING)
-                                            .addComponent(LocalIndexListBox, javax.swing.GroupLayout.Alignment.LEADING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-                                        .addGap(0, 0, Short.MAX_VALUE)))
-                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)))
-                        .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                            .addComponent(jLabel3, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                            .addGroup(layout.createSequentialGroup()
-                                .addGap(24, 24, 24)
-                                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                                    .addComponent(jLabel5)
-                                    .addComponent(ConnectionListBox, javax.swing.GroupLayout.PREFERRED_SIZE, 276, javax.swing.GroupLayout.PREFERRED_SIZE)
-                                    .addComponent(NetworkIndexListBox, javax.swing.GroupLayout.PREFERRED_SIZE, 267, javax.swing.GroupLayout.PREFERRED_SIZE)
-                                    .addComponent(listenConnection))
-                                .addGap(0, 60, Short.MAX_VALUE))))
-                    .addGroup(layout.createSequentialGroup()
-                        .addGap(189, 189, 189)
-                        .addComponent(jLabel2)
-                        .addGap(0, 0, Short.MAX_VALUE)))
-                .addContainerGap())
-            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, layout.createSequentialGroup()
-                .addGap(0, 0, Short.MAX_VALUE)
-                .addComponent(filler1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(81, 81, 81))
-        );
-        layout.setVerticalGroup(
-            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, layout.createSequentialGroup()
-                .addContainerGap()
-                .addComponent(jLabel2)
-                .addGap(31, 31, 31)
-                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(jLabel1)
-                    .addComponent(jLabel3))
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
-                    .addComponent(ConnectionListBox, javax.swing.GroupLayout.DEFAULT_SIZE, 199, Short.MAX_VALUE)
-                    .addComponent(IpListBox, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(ScanNetworkButton)
-                    .addComponent(ConnectButton)
-                    .addComponent(listenConnection))
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 32, Short.MAX_VALUE)
-                .addComponent(filler1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(jLabel4)
-                    .addComponent(jLabel5))
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
-                    .addComponent(LocalIndexListBox, javax.swing.GroupLayout.DEFAULT_SIZE, 195, Short.MAX_VALUE)
-                    .addComponent(NetworkIndexListBox, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-                .addGap(44, 44, 44))
-        );
-
-        pack();
-    }// </editor-fold>//GEN-END:initComponents
-
-    private void ScanNetworkButtonActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_ScanNetworkButtonActionPerformed
-        Executors.newSingleThreadExecutor().execute(new Runnable() {
-            @Override
-            public void run() {
-                try {
-                    IpListBox.removeAll();
-                    //FIXME: Up Iplerin gelmesi uzun suruyor. Ve sonradan gelenler oluyor ??!
-                    mainIntroducer = gabriel.getMainIntroducer();
-                    mainIntroducer.setNodes(new ArrayList<>());
-                    mainIntroducer.checkHostsBruteForce(gabriel.getSubnet());
-                    Thread.sleep(1);
-                    mainIntroducer.getHostCheckerAll().getHostIps().forEach((node) -> {
-                        IpListBox.add(node);
-                    });
-                } catch (InterruptedException ex) {
-                    Logger.getLogger(MainScreen.class.getName()).log(Level.SEVERE, null, ex);
-                }
-            }
-        });
-
-    }//GEN-LAST:event_ScanNetworkButtonActionPerformed
-
-    private void ConnectButtonActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_ConnectButtonActionPerformed
-        String selectedIp = IpListBox.getSelectedItem();
-        if (selectedIp == null) {
-            JOptionPane.showMessageDialog(null, "Please choose 1 IP to connect to.");
-        } else {
-            try {
-                String ip = selectedIp.split(" -")[0].replaceAll("/", "");
-                Node selectedNode = gabriel.getNode(ip);
-                Connection newConnection = new Connection((Inet4Address) selectedNode.getConnectionIp());
-
-                connectionManager.addConnection(newConnection);
-                if (connectionManager.startConnection(newConnection)) {
-                    ConnectionListBox.add(newConnection.getConnectionIp().toString());
-                } else {
-                    System.out.println("Sorry, could not connect to: " + newConnection.getConnectionIp().toString());
-                }
-            } catch (HeadlessException ex) {
-                Logger.getLogger(MainScreen.class.getName()).log(Level.SEVERE, null, ex);
-            }
-        }
-    }//GEN-LAST:event_ConnectButtonActionPerformed
-
-    private void IpListBoxActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_IpListBoxActionPerformed
-        // TODO add your handling code here:
-    }//GEN-LAST:event_IpListBoxActionPerformed
-
-    private void listenConnectionActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_listenConnectionActionPerformed
-
-        try {
-            //Listening for a connection to be made            
-            ServerSocket serverSocket = new ServerSocket(connectionManager.getPORT_NO());
-            System.out.println("TCPServer Waiting for client on port 5000\n");
-            Socket connectionSocket = serverSocket.accept();
-
-            Connection newIncomingConnection = new Connection((Inet4Address) connectionSocket.getInetAddress());
-            connectionManager.addConnection(newIncomingConnection);
-            Index nodeIndex =connectionManager.getIndex(connectionSocket);
-            
-            connectionManager.addToNetworkIndex(nodeIndex);
-            
-//          connectionManager.networkIndex.getFileHeaders().addAll(incomingData.getFileHeaders());
-            System.out.println(nodeIndex.getFileHeaders().get(0).getName());
-            
-            
-            ConnectionListBox.add(newIncomingConnection.getConnectionIp().toString());
-
-        } catch (IOException ex) {
-            System.out.println("Input Output Exception on Listen Connection Action Performed");
-            Logger.getLogger(MainScreen.class.getName()).log(Level.SEVERE, null, ex);
-        }
-    }//GEN-LAST:event_listenConnectionActionPerformed
-
-    /**
-     * @param args the command line arguments
-     */
-    public static void main(String args[]) throws InterruptedException {
-        /* Set the Nimbus look and feel */
-        //<editor-fold defaultstate="collapsed" desc=" Look and feel setting code (optional) ">
-        /* If Nimbus (introduced in Java SE 6) is not available, stay with the default look and feel.
-         * For details see http://download.oracle.com/javase/tutorial/uiswing/lookandfeel/plaf.html 
-         */
-        try {
-            for (javax.swing.UIManager.LookAndFeelInfo info : javax.swing.UIManager.getInstalledLookAndFeels()) {
-                if ("Nimbus".equals(info.getName())) {
-                    javax.swing.UIManager.setLookAndFeel(info.getClassName());
-                    break;
-                }
-            }
-        } catch (ClassNotFoundException ex) {
-            java.util.logging.Logger.getLogger(MainScreen.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (InstantiationException ex) {
-            java.util.logging.Logger.getLogger(MainScreen.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (IllegalAccessException ex) {
-            java.util.logging.Logger.getLogger(MainScreen.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (javax.swing.UnsupportedLookAndFeelException ex) {
-            java.util.logging.Logger.getLogger(MainScreen.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        }
-        //</editor-fold>
-
-        /* Create and display the form */
-        java.awt.EventQueue.invokeLater(new Runnable() {
-            public void run() {
-                new MainScreen().setVisible(true);
-            }
-        });
-
-    }
-
-
-    // Variables declaration - do not modify//GEN-BEGIN:variables
-    private javax.swing.JButton ConnectButton;
-    private java.awt.List ConnectionListBox;
-    private java.awt.List IpListBox;
-    private java.awt.List LocalIndexListBox;
-    private java.awt.List NetworkIndexListBox;
-    private javax.swing.JButton ScanNetworkButton;
-    private javax.swing.Box.Filler filler1;
-    private javax.swing.JLabel jLabel1;
-    private javax.swing.JLabel jLabel2;
-    private javax.swing.JLabel jLabel3;
-    private javax.swing.JLabel jLabel4;
-    private javax.swing.JLabel jLabel5;
-    private javax.swing.JButton listenConnection;
-    // End of variables declaration//GEN-END:variables
-}
