--- src/gabriel/models/Host.java
@@ -1,28 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package gabriel.models;
-
-/**
- *
- * @author TheDoctor
- */
-public class Host {
-
-    String hostIp;
-
-    public Host(String hostIp) {
-        this.hostIp = hostIp;
-    }
-
-    public String getHostIp() {
-        return hostIp;
-    }
-
-    public void setHostIp(String hostIp) {
-        this.hostIp = hostIp;
-    }
-
-}
