+++ src/main/java/org/organet/commons/gabriel/Controller/Hasher.java
@@ -1,0 +1,10 @@
+package org.organet.commons.gabriel.Controller;
+
+import java.io.IOException;
+import java.security.NoSuchAlgorithmException;
+
+public class Hasher {
+  public static String calculateFileHash(String path) throws IOException, NoSuchAlgorithmException {
+    throw new UnsupportedOperationException("Not supported yet."); // FIXME
+  }
+}
