+++ src/main/java/org/organet/commons/gabriel/Controller/Receiver.java
@@ -1,0 +1,11 @@
+package org.organet.commons.gabriel.Controller;
+
+/**
+ * This class is to Receive Messages and Resources from other Connections
+ */
+public class Receiver implements Runnable{
+  @Override
+  public void run() {
+    throw new UnsupportedOperationException("Not supported yet."); // FIXME
+  }
+}
