+++ src/main/java/org/organet/commons/gabriel/Controller/Sender.java
@@ -1,0 +1,11 @@
+package org.organet.commons.gabriel.Controller;
+
+/**
+ * This class is to send Messages and Resources
+ */
+public class Sender implements Runnable {
+  @Override
+  public void run() {
+    throw new UnsupportedOperationException("Not supported yet."); // FIXME
+  }
+}
