+++ src/main/java/org/organet/commons/gabriel/Model/Host.java
@@ -1,0 +1,17 @@
+package org.organet.commons.gabriel.Model;
+
+public class Host {
+  String hostIp;
+
+  public Host(String hostIp) {
+    this.hostIp = hostIp;
+  }
+
+  public String getHostIp() {
+    return hostIp;
+  }
+
+  public void setHostIp(String hostIp) {
+    this.hostIp = hostIp;
+  }
+}
